DCS Codex App Server/requirements.txt
appnope==0.1.0
asgiref==3.2.3
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cffi==1.12.3
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==3.0.2
djangorestframework==3.11.0
entrypoints==0.3
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
powerline-status==2.7
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.7
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
scipy==1.3.3
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
